Split up tcp/tls/local sessions
